Vulnerability adapters are extracted from use_searbar_configuration
